Final round of Wall fixes for today. If you see any build problem - please contact me. It works w/o warnings with build_all on pcamsdt1 - but that's only one machine :-) Different compilers may exhibit different warnings/error.
This commit enables Wall for gcc & icc - please report any warnings that are still leftover, with following exceptions:

- Wno-unknown-pragmas is enabled for gcc (until the OpenMP build is fixed)
- Wno-unused-function (cfortran.h defines static functions -- needs to be moved into cfortran.c so we can disable this suppression)
- Wno-write-strings (lots of char* -> const char* conversions need to be applied - Matthew is already working on that)
- Wno-missing-braces (requires boring work in the Tof* classes, which define multi-dimensional arrays w/o any braces except the outer ones... I fixed most, except these).
- Wno-strict-aliasing (to be investigated)

Please note that Werror is NOT enabled (which would turn all warnings into errors) - we should do this after a grace period, when there are
no more warnings left. I don't want to break other peoples builds with cleanups like this.

+++ CC/TofSimUtil.C
@@ -649,13 +649,13 @@ void TofSimUtil::DefineTOF_PME(){
   if (TOFPM_Et)return;
   TOFPM_Et= new G4PhysicsTable();
   G4PhysicsOrderedFreeVector* PMTEnVector=new G4PhysicsOrderedFreeVector();
-  for (size_t i=0; i<ENUM; i++){
+  for (int i=0; i<ENUM; i++){
      PMTEnVector->InsertValues(PhEnergy[i], EnEFF[i]);
    }
    TOFPM_Et->insertAt(0,PMTEnVector);
 
   G4PhysicsOrderedFreeVector* PMTAnVector=new G4PhysicsOrderedFreeVector();
-  for (size_t i=0; i<ANUM; i++){
+  for (int i=0; i<ANUM; i++){
      PMTAnVector->InsertValues(CosAn[i], AnEFF[i]);
    }
    TOFPM_Et->insertAt(1,PMTAnVector);
